requirements.txt
# ============================================
# CORE DATA SCIENCE & NUMERICAL COMPUTING
# ============================================
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1

# ============================================
# MACHINE LEARNING
# ============================================
scikit-learn==1.6.1

# ============================================
# DEEP LEARNING FRAMEWORKS
# ============================================
torch==2.6.0
torchvision==0.21.0
tensorflow==2.18.0
tf-keras 

# ============================================
# TRANSFORMERS & NLP
# ============================================
transformers==4.45.2
sentence-transformers==3.3.1
huggingface-hub==0.34.0
tokenizers==0.20.0

# ============================================
# VECTOR DATABASES
# ============================================
chromadb==0.5.23
qdrant-client==1.12.1

# ============================================
# LLM & AI FRAMEWORKS
# ============================================
openai==1.58.1
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.14
llama-index==0.12.11
langchain-google-genai==2.0.8
langchain-groq==0.2.0
# Replaced deprecated google-generativeai with google-genai
google-genai
# Explicitly required for text splitting import
langchain-text-splitters
langsmith==0.1.147
# ============================================
# EMBEDDING MODELS
# ============================================
faiss-cpu==1.9.0.post1

# ============================================
# DATA VISUALIZATION
# ============================================
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1

# ============================================
# WEB APPLICATIONS & UI
# ============================================
streamlit==1.52.2
gradio==5.9.1
flask==3.1.0
fastapi==0.115.6
uvicorn==0.34.0
watchfiles==1.0.3

# ============================================
# JUPYTER & NOTEBOOKS
# ============================================
jupyter==1.1.1
jupyterlab==4.3.4
ipykernel==6.30.0
notebook==7.3.2

# ============================================
# COMPUTER VISION
# ============================================
opencv-python==4.10.0.84
Pillow==11.0.0

# ============================================
# NLP LIBRARIES
# ============================================
nltk==3.9.1
spacy==3.8.3

# ============================================
# WEB SCRAPING & APIs
# ============================================
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.28.1

# ============================================
# DATABASE CLIENTS
# ============================================
pymongo==4.10.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
redis==5.2.1

# ============================================
# DATA PROCESSING
# ============================================
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.67.1
python-multipart==0.0.20

# ============================================
# PDF & DOCUMENT PROCESSING
# ============================================
pypdf==5.1.0
python-docx==1.1.2

# ============================================
# UTILITIES
# ============================================
python-dateutil==2.9.0
pydantic==2.10.4
pydantic-settings==2.7.1